Getting ready to add doctor page

+++ src/lib/services/secure_storage.dart
@@ -3,16 +3,21 @@ import "package:flutter_secure_storage/flutter_secure_storage.dart";
 class SecureStorage {
   static final storage = FlutterSecureStorage();
 
-  static Future<void> write(String key, String value) async {
+  static Future<void> writeOne(String key, String value) async {
     await storage.write(key: key, value: value);
   }
 
-  static Future<String?> read(String key) async {
-    String? value = await storage.read(key: key);
-    return value;
+  static Future<void> writeMany(Map<String, String> pairs) async {
+    pairs.forEach((key, value) async => await storage.write(key: key, value: value));
   }
 
-  static Future<void> delete(String key) async {
-    await storage.delete(key: key);
+  static Future<dynamic> read([String? key]) async {
+    if (key != null) return await storage.read(key: key);
+    return await storage.readAll();
+  }
+
+  static Future<void> delete([String? key]) async {
+    if (key != null) await storage.delete(key: key);
+    else await storage.deleteAll();
   }
 }
